+++ myia/compile/__init__.py
@@ -1,4 +1,3 @@
 """Compilation of graphs into optimized code."""
 
-from .vm import FinalVM  # noqa
-from .step import step_compile  # noqa
+from .backends import load_backend  # noqa
