+++ myia/pipeline/steps.py
@@ -19,6 +19,7 @@ from ..utils import overload, flatten, no_prof
 from ..validate import validate, whitelist as default_whitelist, \
     validate_abstract as default_validate_abstract
 from ..vm import VM
+from .compile.backends import load_backend
 
 from .pipeline import pipeline_function, PipelineStep
 
@@ -429,6 +430,42 @@ def step_cconv(self, graph):
     """
     closure_convert(graph)
     return {'graph': graph}
+
+
+###############
+# Compilation #
+###############
+
+
+class CompileStep(PipelineStep):
+    """Step to compile a graph to a configurable backend.
+
+    Inputs:
+        graph: a graph (must be typed)
+
+    Outputs:
+        output: a callable
+
+    """
+    def __init__(self, pipeline_init, backend=None, backend_options=None):
+        """Initialize a CompileStep.
+
+        Arguments:
+            backend: (str) the name of the backend to use
+            backend_options: (dict) options for the backend
+
+        """
+        super().__init__(pipeline_init)
+        backend_class = load_backend(backend)
+        self.backend = backend_class(**backend_options)
+
+    def step(self, graph):
+        """Compile the set of graphs."""
+        out = self.backend.compile(graph)
+        return {'output': out}
+
+
+step_compile = CompileStep.partial()
 
 
 ############################
